% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fastgbm.R
\name{coef.fastgbm}
\alias{coef.fastgbm}
\title{Tree ensembles have no ordinary regression coefficients}
\usage{
\method{coef}{fastgbm}(object, ...)
}
\arguments{
\item{object}{A fitted `fastgbm` object.}

\item{...}{Unused.}
}
\value{
Always errors.
}
\description{
Tree ensembles have no ordinary regression coefficients
}
