% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.survgbm.R
\name{coef.survgbm}
\alias{coef.survgbm}
\title{Tree ensembles have no ordinary regression coefficients}
\usage{
\method{coef}{survgbm}(object, ...)
}
\arguments{
\item{object}{A fitted `survgbm` object.}

\item{...}{Unused.}
}
\value{
Always errors.
}
\description{
Tree ensembles have no ordinary regression coefficients
}
