% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{load_fastgbm}
\alias{load_fastgbm}
\title{Load a serialized fastgbm model}
\usage{
load_fastgbm(path)
}
\arguments{
\item{path}{Path to a serialized model, as written by [save_fastgbm()].}
}
\value{
A `fastgbm` object.
}
\description{
Load a serialized fastgbm model
}
