% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{load_survgbm}
\alias{load_survgbm}
\title{Load a serialized survgbm model}
\usage{
load_survgbm(path)
}
\arguments{
\item{path}{Path to a serialized model, as written by [save_survgbm()].}
}
\value{
A `survgbm` object.
}
\description{
Load a serialized survgbm model
}
