% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fastgbm.R
\name{print.fastgbm}
\alias{print.fastgbm}
\title{Print a fitted fastgbm model}
\usage{
\method{print}{fastgbm}(x, ...)
}
\arguments{
\item{x}{A fitted `fastgbm` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a fitted fastgbm model
}
