% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgbm_multiclass.R
\name{print.fastgbm_multiclass}
\alias{print.fastgbm_multiclass}
\title{Print a fitted multiclass fastgbm model}
\usage{
\method{print}{fastgbm_multiclass}(x, ...)
}
\arguments{
\item{x}{A fitted `fastgbm_multiclass` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a fitted multiclass fastgbm model
}
