% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fastgbm.R
\name{print.summary.fastgbm}
\alias{print.summary.fastgbm}
\title{Print a fastgbm model summary}
\usage{
\method{print}{summary.fastgbm}(x, ...)
}
\arguments{
\item{x}{A `summary.fastgbm` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a fastgbm model summary
}
