% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survgbm.R
\name{print.summary.survgbm}
\alias{print.summary.survgbm}
\title{Print a survgbm model summary}
\usage{
\method{print}{summary.survgbm}(x, ...)
}
\arguments{
\item{x}{A `summary.survgbm` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a survgbm model summary
}
