% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survgbm.R
\name{print.survgbm}
\alias{print.survgbm}
\title{Print a fitted survgbm model}
\usage{
\method{print}{survgbm}(x, ...)
}
\arguments{
\item{x}{A fitted `survgbm` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a fitted survgbm model
}
