% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{save_fastgbm}
\alias{save_fastgbm}
\title{Save a fitted fastgbm model}
\usage{
save_fastgbm(object, path)
}
\arguments{
\item{object}{A fitted `fastgbm` object.}

\item{path}{Path to write the serialized model to.}
}
\value{
`path`, invisibly.
}
\description{
Save a fitted fastgbm model
}
