% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fastgbm.R
\name{summary.fastgbm}
\alias{summary.fastgbm}
\title{Summarize a fitted fastgbm model}
\usage{
\method{summary}{fastgbm}(object, ...)
}
\arguments{
\item{object}{A fitted `fastgbm` object.}

\item{...}{Unused.}
}
\value{
A `summary.fastgbm` object.
}
\description{
Summarize a fitted fastgbm model
}
