% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survgbm.R
\name{summary.survgbm}
\alias{summary.survgbm}
\title{Summarize a fitted survgbm model}
\usage{
\method{summary}{survgbm}(object, ...)
}
\arguments{
\item{object}{A fitted `survgbm` object.}

\item{...}{Unused.}
}
\value{
A `summary.survgbm` object.
}
\description{
Summarize a fitted survgbm model
}
