% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survgbm-package.R
\docType{package}
\name{survgbm-package}
\alias{survgbm-package}
\title{survgbm: Compact Gradient Boosting for Survival Analysis}
\description{
A compact, survival-analysis-only gradient boosting machine with a
compiled (Rcpp + RcppParallel) backend.
}
\author{
\strong{Maintainer}: Imad El-Badisy \email{elbadisyimad@gmail.com}

}
\keyword{internal}
