% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgbm-package.R
\docType{package}
\name{fastgbm-package}
\alias{fastgbm-package}
\title{fastgbm: Compact Gradient Boosting for Regression, Classification, and Survival Analysis}
\description{
A compact gradient boosting machine with a compiled (Rcpp + RcppParallel)
backend, covering regression, binary classification, and right-censored
survival analysis with one interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ielbadisy/fastgbm}
  \item Report bugs at \url{https://github.com/ielbadisy/fastgbm/issues}
}

}
\author{
\strong{Maintainer}: Imad EL BADISY \email{elbadisyimad@gmail.com}

}
\keyword{internal}
